+++ packages/storefront/packages/nuxt/CHANGELOG.md
@@ -1,5 +1,42 @@
 # Change log
 
+## 3.1.0
+
+- **[ADDED]** `middlewareModule` to `defineSdkConfig` params.
+
+```diff [sdk.config.ts]
+- import { UnifiedApiExtension } from "storefront-middleware/types"
++ import { UnifiedEndpoints } from "storefront-middleware/types"
+
+export default defineSdkConfig(
+-  ({ buildModule, middlewareUrl, getRequestHeaders }) => ({
+-    commerce: buildModule(unifiedModule<UnifiedApiExtension>, {
+-      apiUrl: `${middlewareUrl}/commerce`,
+-      requestOptions: { headers: getRequestHeaders },
++  ({ buildModule, middlewareModule, middlewareUrl, getRequestHeaders }) => ({
++      commerce: buildModule(middlewareModule<UnifiedEndpoints>, {
++        apiUrl: `${middlewareUrl}/commerce`,
++        defaultRequestConfig: { headers: getRequestHeaders() },
+      }),
+  })
+);
+```
+
+- **[CHANGED]** deprecate `getCookieHeader`, use `getRequestHeaders` instead
+
+```diff [sdk.config.ts]
+export default defineSdkConfig(
+-  ({ buildModule, middlewareModule, middlewareUrl, getCookieHeader }) => ({
++  ({ buildModule, middlewareModule, middlewareUrl, getRequestHeaders }) => ({
+      commerce: buildModule(middlewareModule<UnifiedEndpoints>, {
+        apiUrl: `${middlewareUrl}/commerce`,
+-        defaultRequestConfig: { headers: getCookieHeader() }, // Only cookie header is included.
++        defaultRequestConfig: { headers: getRequestHeaders() }, // All headers are included.
+      }),
+  })
+);
+```
+
 ## 3.0.3
 
 - **[CHANGED]** `@nuxt/kit` locked `3.7.4` version to `@nuxt/kit@^3.7.4`